API/requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
django-cors-headers==4.0.0
Pillow~=9.5.0
